+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `network set_max_communities <network> <max_communities>` command to set the maximum number of communities allowed on a network.
 - `network unset_max_communities <network>` command to unset the maximum number of communities allowed on a network.
 
+### Changed
+
+- Renamed `network policy_set_prefix` to `network policy_set_pattern`.
+- Renamed `network policy_unset_prefix` to `network policy_unset_pattern`.
+
 ## [1.7.3](https://github.com/unioslo/mreg-cli/releases/tag/1.7.3) - 2025-11-18
 
 ### Changed