+++ CHANGELOG.md
@@ -23,7 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- `logging start` showing "disabled" instead of the actual logging status."
+- `logging start` showing "disabled" instead of the actual logging status.
 - `host remove` error message suggesting incorrect option and syntax when `-override` is required.
 
 ## [1.7.3](https://github.com/unioslo/mreg-cli/releases/tag/1.7.3) - 2025-11-18